Move some dialogs
