+++ source/renderer/App.tsx
@@ -16,8 +16,8 @@ import { PreferencesDialog } from "./components/PreferencesDialog";
 import { Notifications } from "./components/Notifications";
 import { FileHostConnectionNotice } from "./components/FileHostConnectionNotice";
 import { CreateNewFilePrompt } from "./components/CreateNewFilePrompt";
-import { UpdateDialog } from "./components/UpdateDialog";
-import { AboutDialog } from "./components/AboutDialog";
+import { UpdateDialog } from "./components/standalone/UpdateDialog";
+import { AboutDialog } from "./components/standalone/AboutDialog";
 import { BiometricRegistrationDialog } from "./components/BiometricRegistrationDialog";
 import { useTheme } from "./hooks/theme";
 import { Theme } from "./types";
